Storage Integration Tests (#112)

+++ modyn/storage/internal/database/models/file.py
@@ -2,7 +2,7 @@
 
 from modyn.storage.internal.database.models.dataset import Dataset
 from modyn.storage.internal.database.storage_base import Base
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, ForeignKey, Integer, String
 from sqlalchemy.orm import backref, relationship
 
 
@@ -14,8 +14,8 @@ class File(Base):
     dataset_id = Column(Integer, ForeignKey("datasets.dataset_id"), nullable=False)
     dataset = relationship("Dataset", backref=backref("files", lazy=True))
     path = Column(String(120), unique=False, nullable=False)
-    created_at = Column(Integer, nullable=False)
-    updated_at = Column(Integer, nullable=False)
+    created_at = Column(BigInteger, nullable=False)
+    updated_at = Column(BigInteger, nullable=False)
     number_of_samples = Column(Integer, nullable=False)
 
     def __repr__(self) -> str:
